+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/BuiltinCallable.java
@@ -1,10 +1,16 @@
 package com.revolvingmadness.sculk.language.builtins.classes;
 
 import com.revolvingmadness.sculk.language.ErrorHolder;
+import com.revolvingmadness.sculk.language.builtins.classes.types.CallableType;
 
 import java.util.List;
 
 public abstract class BuiltinCallable extends BuiltinClass {
+    @Override
+    public BuiltinType getType() {
+        return CallableType.TYPE;
+    }
+
     public void validate(String callableName, List<BuiltinClass> arguments, List<BuiltinType> argumentTypes) {
         if (arguments.size() != argumentTypes.size()) {
             throw ErrorHolder.invalidArgumentCount(callableName, argumentTypes.size(), arguments.size());
@@ -18,6 +24,8 @@ public abstract class BuiltinCallable extends BuiltinClass {
             if (!argument.instanceOf(argumentType)) {
                 throw ErrorHolder.argumentRequiresType(argumentNumber + 1, callableName, argument.getType(), argumentType);
             }
+
+            argumentNumber++;
         }
     }