+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/types/PlayerEntityType.java
@@ -1,15 +1,15 @@
 package com.revolvingmadness.sculk.language.builtins.classes.types;
 
+import com.revolvingmadness.sculk.language.GUIScreenHandler;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinClass;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinMethod;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinType;
-import com.revolvingmadness.sculk.language.builtins.classes.instances.BooleanInstance;
-import com.revolvingmadness.sculk.language.builtins.classes.instances.NullInstance;
-import com.revolvingmadness.sculk.language.builtins.classes.instances.StringInstance;
-import com.revolvingmadness.sculk.language.builtins.classes.instances.WorldInstance;
+import com.revolvingmadness.sculk.language.builtins.classes.instances.*;
 import com.revolvingmadness.sculk.language.interpreter.Interpreter;
 import com.revolvingmadness.sculk.language.lexer.TokenType;
+import net.minecraft.screen.SimpleNamedScreenHandlerFactory;
 import net.minecraft.server.world.ServerWorld;
+import net.minecraft.text.Text;
 import net.minecraft.world.World;
 
 import java.util.List;
@@ -27,6 +27,9 @@ public class PlayerEntityType extends BuiltinType {
         this.typeVariableScope.declare(List.of(TokenType.CONST), "getName", new GetName());
         this.typeVariableScope.declare(List.of(TokenType.CONST), "getUUID", new GetUUID());
         this.typeVariableScope.declare(List.of(TokenType.CONST), "getWorld", new GetWorld());
+        this.typeVariableScope.declare(List.of(TokenType.CONST), "openGUI", new OpenGUI());
+        this.typeVariableScope.declare(List.of(TokenType.CONST), "getStackInMainHand", new GetStackInMainHand());
+        this.typeVariableScope.declare(List.of(TokenType.CONST), "getStackInOffHand", new GetStackInOffHand());
     }
 
     private static class AddExperienceLevels extends BuiltinMethod {
@@ -61,6 +64,24 @@ public class PlayerEntityType extends BuiltinType {
             this.validate("getName", arguments);
 
             return new StringInstance(this.boundClass.toPlayerEntity().getName().getLiteralString());
+        }
+    }
+
+    private static class GetStackInMainHand extends BuiltinMethod {
+        @Override
+        public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
+            this.validate("getStackInMainHand", arguments, List.of());
+
+            return new ItemStackInstance(this.boundClass.toPlayerEntity().getMainHandStack());
+        }
+    }
+
+    private static class GetStackInOffHand extends BuiltinMethod {
+        @Override
+        public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
+            this.validate("getStackInOffHand", arguments, List.of());
+
+            return new ItemStackInstance(this.boundClass.toPlayerEntity().getOffHandStack());
         }
     }
 
@@ -103,6 +124,19 @@ public class PlayerEntityType extends BuiltinType {
             this.validate("isSpectator", arguments);
 
             return new BooleanInstance(this.boundClass.toPlayerEntity().isSpectator());
+        }
+    }
+
+    private static class OpenGUI extends BuiltinMethod {
+        @Override
+        public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
+            this.validate("openGUI", arguments, List.of(GUIType.TYPE));
+
+            GUIInstance gui = arguments.get(0).toGUI();
+
+            this.boundClass.toPlayerEntity().openHandledScreen(new SimpleNamedScreenHandlerFactory((syncId, playerInventory, player) -> new GUIScreenHandler(interpreter, gui, syncId, playerInventory, gui.inventory), Text.literal(gui.title)));
+
+            return new NullInstance();
         }
     }
 }